Refactor hiera hierarchy and create buildfarm_role.
This uses hiera to keep things a little cleaner. It uses an actually
common common.yaml for things like the master and repo IPs and other
shared values, then uses a custom fact, buildfarm_role, to apply
hierarchy specific to that role.

I have not yet changed the puppet invocation as I also need to create
the role modules to complement the profile ones.