Refactor hiera and role specification.
With this change and the analogous one in the buildfarm_deployment
repository the different high level roles are refactored from separate
hiera configs and site manifests to a unified hiera hierarchy and role
modules.

The common.yaml hiera data is now actually common and at the base of the
hierarchy. Now there is an extra level: buildfarm_role which is provided
as a custom fact via the `FACTER_buildfarm_role` environment variable.
This is one of agent, master, repo; analogous to the master, repo, slave
directories containing hiera data and a site manifest.