Docker module does not need to install additional kernel packages.
Since the deployment of the new AWS-optimised kernel for Ubuntu in April
2017, buildfarm deployments on AWS will fail unless the operator first
reverts to the default virtual kernel.

The "manage_kernel" parameter is meant to install supplementary kernel
packages required by older combinations of docker and Ubuntu.
https://github.com/garethr/garethr-docker/blob/ccade80caf3be669241e733d6489828789cba5c1/manifests/install.pp#L65-L67

According to the official puppet-docker module, which was copied from
garethr/docker without commit history, this flag is no longer needed any
of their supported platforms [1] and is due to be removed.

Since Ubuntu Xenial is one of the platforms that this is no longer
needed on. Removing it to allow the use of the AWS-optimized kernel does
no harm.

[1]: https://github.com/puppetlabs/puppetlabs-docker/issues/38